+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added `DateSelector` widget component
 
 ## [0.13.0]
-- Renamed `Chat` to `OAIChat`
 
 ## [0.12.1]
 - Fixed `AppBar` portal to inherit font variables from the current `Surface`